Hazard Based Distributions

Custom distribution type

JointModels.HazardBasedDistribution — Type

HazardBasedDistribution

HazardBasedDistribution is a type that builds a distribution based on a hazard function.

To use this type to formulate a distribution yourself: implement a struct and the hazard function.

struct LogLogistic <: HazardBasedDistribution
    α::Real
    β::Real
end

function JointModels.hazard(dist::LogLogistic, t::Real)
    α, β  = dist.α, dist.β
    return ((β / α) * (t / α) ^ (β - 1)) / (1 + (t / α) ^ β)
end

HazardBasedDistribution implements numeric integration to calculate the cumulative hazard and builds a distribution based it. To generate samples it solves an ODE and applies inverse transform sampling.

The Distributions.jl functionalities implemented for HazardBasedDistribution.

JointModels.hazard — Method

represents $h(t)$ needs to be implemented for any struct that subtypes HazardBasedDistribution

JointModels.cumulative_hazard — Method

calculates $H(t) = \int_0^t h(u) \; du$ numerically with a Gauss-Konrad procedure.

Base.rand — Method

Generate a random sample $t \sim \text{dist}$ via inverse transform sampling.

Distributions.ccdf — Method

Calculation of the ccdf / survival function at time $t$ based on the cumulative hazard

$S(t) = \exp(-H(t)) = \exp(-\int h(u) du)$

Distributions.logpdf — Method

Calculation of the log pdf function at time $t$ based on the cumulative hazard

$\log (f(t)) = \log(h(t)\cdot S(t)) = \log( h(t)) - H(t)$

Distributions.pdf — Method

Calculation of the pdf function at time $t$ based on the log pdf.